pyproject.toml
[tool.poetry]
name = "cvetrends"
version = "0.0.1"
description = "cvet is a Python utility for pulling actionable vulnerabilities from cvetrends.com"
authors = ["Nicholas <nanastasi@sprocketsecurity.com>"]
readme = "README.md"
homepage = "https://github.com/puzzlepeaches/cvetrends"
repository = "https://github.com/puzzlepeaches/cvetrends"
exclude = ["tests"]
include = ["LICENSE", "CHANGELOG.md"]
license = "BSD 3-Clause License"
classifiers = [
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
]

packages = [
    { include = "cvetrends"}
]

[tool.poetry.scripts]
cvet = 'cvetrends.__main__:main'

[tool.poetry.dependencies]
python = "^3.8"
click = "*"
click-config-file = "*"
rich = "^12.4.1"
requests = "^2.27.1"
notifiers = "^1.3.3"
python-box = "^6.0.2"
rich-click = "^1.5.2"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
black = "*"
pre-commit = "*"
isort = "*"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
